Added visualizations for co2 ground truth and changed data structure …

+++ scripts/download_capno.sh
@@ -1,9 +1,16 @@
+if [ -e data/capno ]; then
+    echo "Error! Dataset already already exists @ data/capno folder, please delete it first."
+    echo " Exiting Script..."
+    return
+fi
 wget "http://www.capnobase.org/uploads/media/TBME2013-PPGRR-Benchmark_R3.zip"
 mkdir data
-mkdir data/raw
-mv TBME2013-PPGRR-Benchmark_R3.zip ./data/raw
-cd ./data/raw
+mv TBME2013-PPGRR-Benchmark_R3.zip ./data
+cd ./data
+mkdir capno
 unzip TBME2013-PPGRR-Benchmark_R3.zip
-mv TBME2013-PPGRR-Benchmark_R3 capno
+mv TBME2013-PPGRR-Benchmark_R3/data capno/raw
+mv TBME2013-PPGRR-Benchmark_R3/README.txt capno/raw
 rm TBME2013-PPGRR-Benchmark_R3.zip
-cd ../..
+rm TBME2013-PPGRR-Benchmark_R3 -r
+cd ../
